# pet-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pets API",
    "version": "1.0.0",
    "description": "A simple Express Pets API\n\nYou can find the Swagger specification files here:\n- [Swagger JSON](/documentation/swagger.json)\n- [Swagger YAML](/documentation/swagger.yaml)"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "schemas": {
      "Pet": {
        "type": "object",
        "required": [
          "id",
          "name",
          "tag"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "The auto-generated id of the pet"
          },
          "name": {
            "type": "string",
            "description": "The name of the pet"
          },
          "tag": {
            "type": "string",
            "description": "The tag of the pet"
          }
        }
      },
      "NewPet": {
        "type": "object",
        "required": [
          "name",
          "tag"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the pet"
          },
          "tag": {
            "type": "string",
            "description": "The tag of the pet"
          }
        }
      }
    }
  },
  "paths": {
    "/pets": {
      "get": {
        "summary": "Retrieve a list of pets",
        "description": "Returns a list of all pets",
        "tags": [
          "Pets"
        ],
        "responses": {
          "200": {
            "description": "A list of pets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new pet",
        "description": "Adds a new pet to the system",
        "tags": [
          "Pets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPet"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/pets/{petId}": {
      "get": {
        "summary": "Get a pet by ID",
        "description": "Retrieve a pet's details by its ID",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "required": true,
            "description": "Numeric ID of the pet to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a pet",
        "description": "Update a pet's details by its ID",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "required": true,
            "description": "Numeric ID of the pet to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated pet details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a pet",
        "description": "Delete a pet by its ID",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "required": true,
            "description": "Numeric ID of the pet to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pet successfully deleted"
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "tags": []
}