# pet-frontend/src/api/models/newPet.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Pets API
 * A simple Express Pets API

You can find the Swagger specification files here:
- [Swagger JSON](/documentation/swagger.json)
- [Swagger YAML](/documentation/swagger.yaml)
 * OpenAPI spec version: 1.0.0
 */

export interface NewPet {
  /** The name of the pet */
  name: string;
  /** The tag of the pet */
  tag: string;
}